rename geminiUploadToS3 to uploadFileToS3
